+++ packages/sdk-metrics/src/utils.ts
@@ -93,50 +93,6 @@ export function callWithTimeout<T>(
   );
 }
 
-export interface PromiseAllSettledFulfillResult<T> {
-  status: 'fulfilled';
-  value: T;
-}
-
-export interface PromiseAllSettledRejectionResult {
-  status: 'rejected';
-  reason: unknown;
-}
-
-export type PromiseAllSettledResult<T> =
-  | PromiseAllSettledFulfillResult<T>
-  | PromiseAllSettledRejectionResult;
-
-/**
- * Node.js v12.9 lower and browser compatible `Promise.allSettled`.
- */
-export async function PromiseAllSettled<T>(
-  promises: Promise<T>[]
-): Promise<PromiseAllSettledResult<T>[]> {
-  return Promise.all(
-    promises.map<Promise<PromiseAllSettledResult<T>>>(async p => {
-      try {
-        const ret = await p;
-        return {
-          status: 'fulfilled',
-          value: ret,
-        };
-      } catch (e) {
-        return {
-          status: 'rejected',
-          reason: e,
-        };
-      }
-    })
-  );
-}
-
-export function isPromiseAllSettledRejectionResult(
-  it: PromiseAllSettledResult<unknown>
-): it is PromiseAllSettledRejectionResult {
-  return it.status === 'rejected';
-}
-
 export function setEquals(lhs: Set<unknown>, rhs: Set<unknown>): boolean {
   if (lhs.size !== rhs.size) {
     return false;