+++ CHANGELOG.md
@@ -21,6 +21,7 @@ For notes on migrating to 2.x / 0.200.x see [the upgrade guide](doc/upgrade-to-2
 ### :house: Internal
 
 * chore(browser): fix CODEOWNERS paths for browser-related packages
+* refactor(sdk-metrics): remove Promise.allSettled() ponyfill
 
 ## 2.3.0